+++ src/header.js
@@ -1,6 +1,6 @@
 "use strict";
 
-/* jsCore JavaScript library v0.4
+/* jsCore JavaScript library v0.4.1
  * © 2014 Dmitry Korobkin
  * Released under the MIT license
  * https://github.com/Octane/jsCore/